% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Relative effects plot}
\usage{
forest_plot(hpp_mod, mod_type)
}
\arguments{
\item{hpp_mod}{A network meta-analysis model object.}

\item{mod_type}{Type of model.}
}
\description{
Relative effects plot
}
