% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{hpp_app}
\alias{hpp_app}
\title{Shiny application for chronic pain and antidepressants}
\usage{
hpp_app(...)
}
\description{
Display dashboard of network meta-analysis results for treatment of chronic
pain with antidepressants.
}
