% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpp_obs.R
\docType{data}
\name{hpp_obs}
\alias{hpp_obs}
\title{Observation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2067 rows and 48 columns.
}
\usage{
hpp_obs
}
\description{
See \link{update_hpp_dat} for target this is taken from.
}
\keyword{datasets}
