% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpp_outcomes.R
\docType{data}
\name{hpp_outcomes}
\alias{hpp_outcomes}
\title{Outcomes in the analysis}
\format{
An object of class \code{character} of length 9.
}
\usage{
hpp_outcomes
}
\description{
Character vector of outcomes in the order the pipeline processes in.
}
\keyword{datasets}
