% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_key.R
\docType{data}
\name{m_key}
\alias{m_key}
\title{Model key}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
m_key
}
\description{
Table defining if outcome is modelling using \code{lor} for log odds ratio or
\code{smd} for standardised mean difference.
}
\keyword{datasets}
