% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_pain_int.R
\docType{data}
\name{m_pain_int}
\alias{m_pain_int}
\title{Pain intensity model output}
\format{
An object of class \code{stan_nma} of length 10.
}
\usage{
m_pain_int
}
\description{
All in model results; all conditions, all antidepressants.
}
\keyword{datasets}
