% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_tab.R
\name{obs_tab}
\alias{obs_tab}
\title{Observation-level data}
\usage{
obs_tab(obs, mod_type)
}
\arguments{
\item{obs}{Dataset produced by \link{obs_selected}.}

\item{mod_type}{\code{lor} or \code{smd}.

These are the data that are included in the model.}
}
\description{
Observation-level data
}
