% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_net.R
\name{set_net}
\alias{set_net}
\title{Set network for model}
\usage{
set_net(dat, mod_type)
}
\arguments{
\item{dat}{A dataset produced by \link{obs_selected}.}

\item{mod_type}{Type of model: \code{lor} or \code{smd}.}
}
\description{
Set network for model
}
