% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tab.R
\name{ss_tab}
\alias{ss_tab}
\title{Outcome-specific summary}
\usage{
ss_tab(obs)
}
\arguments{
\item{obs}{Selected observations produced by \link{obs_selected}.

Table of selection-specific summary information.}
}
\description{
Outcome-specific summary
}
